Delete all commented code
