--- pkg/config/write.go
@@ -1,48 +1,0 @@
-/*
-Copyright © 2022 SUSE LLC
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package config
-
-/*import (
-	"github.com/rancher/wrangler/pkg/data/convert"
-	"sigs.k8s.io/yaml"
-)
-
-func PrintInstall(cfg Config) ([]byte, error) {
-	if cfg.Elemental.Install.Password != "" {
-		cfg.Elemental.Install.Password = "******"
-	}
-	data, err := convert.EncodeToMap(cfg.Elemental.Install)
-	if err != nil {
-		return nil, err
-	}
-
-	toYAMLKeys(data)
-	return yaml.Marshal(data)
-}
-
-func toYAMLKeys(data map[string]interface{}) {
-	for k, v := range data {
-		if sub, ok := v.(map[string]interface{}); ok {
-			toYAMLKeys(sub)
-		}
-		newK := convert.ToYAMLKey(k)
-		if newK != k {
-			delete(data, k)
-			data[newK] = v
-		}
-	}
-}*/
